Add files via upload
This commit contains a complete code refactoring bringing it in line with today's version of Home Assistant. Due to recent changes to the Lovelace Dashboards, the older method for "turning on" the slide show doesn't work anymore. The `background:` key no longer accepts CSS shorthand and now requires a structured YAML map with sub-keys. Going forward, if you want to inject custom CSS vars into Home Assistant you have to do so by way of the HA Theme feature. This makes it MUCH easier to change user preferences such as turning on and off the double-tap (more new features below). All you do is edit the theme file and reload it and you're good to go. You can also apply a "theme" code block to anywhere Themes are supported. Globally, by User or individual Lovelace dashboards. In the case of HASlideshow, the documentation README will only refer to applying this to Lovelace Dashboards. If you want to apply this Globally or by User, which is beyond the scope of this addon, then you'll have to read up on how to write your own theme. I'm just simply pointing out that because HASlideshow is taking advantage of HA Themes to apply it's custom CSS vars to Lovelace Dashboards, you can apply it in other scopes where themes are supported.

- New Features
* Added a background transition animation.
* Added the ability to turn on and off the double-tap feature.
* Added the ability to set your own page transition timings (default: 5 min) and animation speeds (default: 1 sec) without having to edit the JS file.

- Bug Fixes
* Updated `bs_getBackgroundElement()` to crawl down to the lowest possible element where CSS theme vars are applied.
* Removed the feature that used to pull pictures from Picsum. HA no longer supports it for security reasons.

- OS/Browser compatibility (stuff that's been tested so far)
* Home Assistant Yellow running a Raspberry Pi 5 compute node. (Lovelace Dashboard Host)
* Windows 11 running Edge or Firefox. (Client)
* iPhone running Edge, Firefox or the HA mobile app (Client)
* Raspberry Pi + Pi Touch Display 2 running Chromium in Kiosk mode. (Client)

...beyond that I fail to see how other OSes, browsers etc. will have issues with the CSS code.

I'll be updating the README shortly and looking into how I can adapt this repository for HACS.