Revert Update haslideshow.js
I had to revert my changes to the image paths. I learned the hard way that if you manually create a sub-directory in a HACS installed plugin directory, an update via HACS will delete that directory and thus the uploaded images.

I'll have to update the readme to make sure users understand not to put a `/backgrounds/` directory within the `/www/community/HASlideshow` install location.