New Feature - Custom Image Paths
Since we have to assign a theme template to each dashboard, it makes sense to allow a user to specify an image directory specific to that dashboard. This allows setting up each dashboard independently with different images and as a side effect, different timings for the update interval and transition speeds. I'm doing it this way rather than forcing a user to edit the main JS file to create custom mappings. Specifying it in the template file is more user friendly.